more tests
renamed data.py into problems.py
add tests/bruteforce.py to compute brute force solutions
3 more problems with 2, 9 and 16 solutions
