+++ tests/test_solution_count.py
@@ -3,8 +3,9 @@ import numpy as np
 from exact_cover_impl import get_solution_count
 from exact_cover.io import DTYPE_FOR_ARRAY
 
-from .data import small_trimino_problem
-from .data import detailed_wikipedia_problem
+from .problems import small_trimino_problem
+from .problems import detailed_wikipedia_problem
+from .problems import bruteforce_problem1
 
 def test_solution_count():
     data = np.array([[1, 0, 0], [0, 1, 0], [0, 1, 1], [0, 0, 1]], dtype=DTYPE_FOR_ARRAY)
@@ -31,6 +32,24 @@ def test_small_trimino_problem():
 
 def test_detailed_wikipedia_problem():
     problem = detailed_wikipedia_problem()
+    data = problem['data']
+    result = get_solution_count(data)
+    assert result == problem['solution_count']
+
+def test_bruteforce_problem1():
+    problem = bruteforce_problem1()
+    data = problem['data']
+    result = get_solution_count(data)
+    assert result == problem['solution_count']
+
+def test_bruteforce_problem2():
+    problem = bruteforce_problem2()
+    data = problem['data']
+    result = get_solution_count(data)
+    assert result == problem['solution_count']
+
+def test_bruteforce_problem3():
+    problem = bruteforce_problem3()
     data = problem['data']
     result = get_solution_count(data)
     assert result == problem['solution_count']
